+++ src/components/FormComponents/CertificationInfoForm/CertificationInfoForm.tsx
@@ -11,8 +11,13 @@ import { IRegisterFields } from "../RegistrationForm/RegistrationForm";
 
 import { DeleteIcon } from "@chakra-ui/icons";
 import { toastFail, toastSuccess } from "@nepMeds/components/Toast";
-import { useDeleteCertificateInfo } from "@nepMeds/service/nepmeds-certificate";
+import {
+  useDeleteCertificateFile,
+  useDeleteCertificateInfo,
+} from "@nepMeds/service/nepmeds-certificate";
+import serverErrorResponse from "@nepMeds/service/serverErrorResponse";
 import { getImageUrl } from "@nepMeds/utils/getImageUrl";
+import { AxiosError } from "axios";
 
 export const CertificationInfoForm = ({
   doctorProfileData,
@@ -34,10 +39,13 @@ export const CertificationInfoForm = ({
     control,
     name: "certification",
   });
+  const deleteCertificateFile = useDeleteCertificateFile();
 
   const mappedImageInfo =
     doctorProfileData?.doctor_certification_info.map(e =>
-      e?.certificate_document.map((e: any) => getImageUrl(e.file))
+      e?.certificate_document.map((e: any) => {
+        return { url: getImageUrl(e?.file), id: e?.id };
+      })
     ) ?? [];
 
   const deleteCertificateInfoRegister = useDeleteCertificateInfo();
@@ -61,7 +69,9 @@ export const CertificationInfoForm = ({
   }, [doctorProfileData, getValues]);
 
   const [selectedImages, setSelectedImages] =
-    useState<Array<Array<File | string | null>>>(mappedImageInfo);
+    useState<Array<Array<File | { url: string; id: string } | null>>>(
+      mappedImageInfo
+    );
   const [, setSelectedImagesFile] = useState<Array<Array<File | null>>>([]);
 
   const handleImageChange = async (
@@ -77,7 +87,10 @@ export const CertificationInfoForm = ({
         updatedImages[certificateIndex] = [
           ...(updatedImages[certificateIndex] || []),
         ];
-        updatedImages[certificateIndex][imageIndex] = imageUrl;
+        updatedImages[certificateIndex][imageIndex] = {
+          url: imageUrl,
+          id: "0",
+        };
         return updatedImages;
       });
 
@@ -138,6 +151,15 @@ export const CertificationInfoForm = ({
     });
   };
 
+  const handleDeleteFile = async (id: number) => {
+    try {
+      await deleteCertificateFile.mutateAsync(id);
+    } catch (error) {
+      const err = serverErrorResponse(error as AxiosError);
+      toastFail(err);
+    }
+  };
+
   return (
     <>
       {fields.map((item, index) => {
@@ -164,6 +186,7 @@ export const CertificationInfoForm = ({
                     background="#F9FAFB"
                     academicIndex={index}
                     helperText={false}
+                    deleteFile={handleDeleteFile}
                     {...field} // Pass the `field` props to ensure integration with `react-hook-form`
                   />
                 )}