+++ src/service/nepmeds-experience.ts
@@ -94,3 +94,24 @@ export const getSingleExperienceInfo = async (id: number) => {
   );
   return response.data.data;
 };
+
+//Delete Experience File
+const deleteExperienceFile = async (id: number) => {
+  const response = await HttpClient.delete(
+    api.experience_file_delete + `${id}/`
+  );
+  return response;
+};
+
+export const useDeleteExperienceFile = () => {
+  const queryClient = useQueryClient();
+  return useMutation(
+    ["delete" + api.experience_file_delete],
+    deleteExperienceFile,
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(api.doctor_profile);
+      },
+    }
+  );
+};